requirements.txt
async-timeout==3.0.1
discord.py==1.5.1
emojis==0.6.0
python-dotenv==0.15.0
Werkzeug==1.0.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
gunicorn==20.0.4
honcho==1.0.1

pip~=22.1
wheel~=0.36.2
click~=7.1.2
ipython~=7.21.0
setuptools~=54.1.2
future~=0.18.2
pymongo~=4.1.1
discord~=1.0.1